+++ ProyectoISW2/ProyectoISW2/Controllers/ManualesController.cs
@@ -18,12 +18,22 @@ namespace ProyectoISW2.Controllers
         // GET: Manuales
         public ActionResult Index()
         {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
             return View(db.Manuals.ToList());
         }
 
         // GET: Manuales/Details/5
         public ActionResult Details(int? id)
         {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,6 +49,11 @@ namespace ProyectoISW2.Controllers
         // GET: Manuales/Create
         public ActionResult Create()
         {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
             return View();
         }
 
@@ -49,6 +64,11 @@ namespace ProyectoISW2.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Nombre,Cantidad,Seccion")] Manual manual)
         {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
             if (ModelState.IsValid)
             {
                 manual.Nombre = manual.Nombre.ToUpper();
@@ -64,6 +84,11 @@ namespace ProyectoISW2.Controllers
         // GET: Manuales/Edit/5
         public ActionResult Edit(int? id)
         {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -83,6 +108,11 @@ namespace ProyectoISW2.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Nombre,Cantidad,Seccion")] Manual manual)
         {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(manual).State = EntityState.Modified;
@@ -95,6 +125,11 @@ namespace ProyectoISW2.Controllers
         // GET: Manuales/Delete/5
         public ActionResult Delete(int? id)
         {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -112,6 +147,11 @@ namespace ProyectoISW2.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
             Manual manual = db.Manuals.Find(id);
             db.Manuals.Remove(manual);
             db.SaveChanges();
@@ -125,6 +165,73 @@ namespace ProyectoISW2.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+        public ActionResult Stock(int? id)
+        {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            Manual manual = db.Manuals.Find(id);
+            if (manual == null)
+            {
+                return HttpNotFound();
+            }
+            return View(manual);
+        }
+
+        // POST: Pruebas/Edit/5
+        // Para protegerse de ataques de publicación excesiva, habilite las propiedades específicas a las que desea enlazarse. Para obtener 
+        // más información vea http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+
+        public ActionResult Stock(int? Id, int? Cantidad)
+        {
+
+            if (Session["userId"] == null)
+            {
+                return Redirect("/Usuarios/Login");
+            }
+            ViewBag.ManualIdd = Id;
+            ViewBag.CantidadManualess = Cantidad;
+            if (ModelState.IsValid)
+            {
+                int cantpbd, cantpf;
+
+                var queryp = (from l in db.Manuals
+                              where l.Id == Id
+                              select l);
+
+
+
+
+                if (queryp.Count() > 0 && queryp != null)
+                {
+                    Manual manual = queryp.First();
+
+                    cantpbd = Convert.ToInt32(manual.Cantidad);
+                    cantpf = cantpbd + Convert.ToInt32(Cantidad);
+                    if (Cantidad < 0)
+                    {
+                        ViewBag.Message = "Manuales no suficientes";
+                    }
+                    else
+                    {
+                        manual.Cantidad = cantpf;
+                        db.SaveChanges();
+                        return Redirect("/Inventario/");
+                    }
+
+
+                }
+
+            }
+            return View();
         }
     }
 }
